# Conflicts:
#	erp-server/erp-server-file/src/main/resources/excel/oms/soB2c.xlsx
#	erp-server/erp-server-oms/src/main/java/com/erp/server/oms/controller/api/SkuMappingController.java
#	erp-server/erp-server-oms/src/main/java/com/erp/server/oms/mapper/SkuMappingMapper.java
#	erp-server/erp-server-oms/src/main/java/com/erp/server/oms/service/SkuMappingService.java
#	erp-server/erp-server-oms/src/main/java/com/erp/server/oms/service/impl/SkuMappingServiceImpl.java
#	erp-server/erp-server-wms/src/main/java/com/erp/server/wms/controller/api/FirstMileDeliveryController.java

+++ erp-sdk/erp-sdk-oms-mercado/src/main/java/com/sdk/oms/mercado/dto/mercado/order/SellerBean.java
@@ -13,6 +13,6 @@ public class SellerBean {
      */
 
     @JsonProperty("id")
-    private int fid;
+    private long fid;
 
 }
